h2の追加

+++ src/main/java/com/example/todosample/TodoEntity.java
@@ -1,7 +1,7 @@
 package com.example.todosample;
 
 public class TodoEntity {
-    Long Id;
+    Integer Id;
     String title;
     Boolean completed;
 }
